refactor: move ProgramInfo generation to programs
